feat: refactoring + client declaration in main + add example test with mock
